pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.43",
    "tqdm>=4.67.1",
    "werkzeug>=3.1.3",
]
